Added subclient support.
This allows an application (the client, such as Blender) to authenticate
against different applications (the subclients, such as Pillar) without
passing the user's actual authentication token to the subclient.

Users can request a SubClient-Specific Token (SCST), which can later be
verified by the subclient.

Subclients are identified by their unique subclient ID, which must be
known to both the client and the subclient.It is now possible to
manage tokens for subclients.